Rename back to "fixpoint_sdk"
Rename this SDK package back to "fixpoint_sdk". We are going to use the
name "fixpoint" for our new repo: https://github.com/gofixpoint/fixpoint/
