+++ bin/pip-freeze
@@ -20,7 +20,7 @@ fi
 
 cd $ROOT_DIR
 
-# Ignore the `fixpoint` if we installed it in editable mode, and ignore
+# Ignore the `fixpoint_sdk` if we installed it in editable mode, and ignore
 # `pkg_resources`, which is included because of some weird bug.
 # https://stackoverflow.com/questions/39577984/what-is-pkg-resources-0-0-0-in-output-of-pip-freeze-command
-pip freeze | grep -E -v '(fixpoint|pkg_resources)' > requirements.txt
+pip freeze | grep -E -v '(fixpoint_sdk|pkg_resources)' > requirements.txt
