+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "fixpoint"
+name = "fixpoint_sdk"
 version = "0.10.0"
 
 authors = [
